INT32 typed parameter fix, Flight time fix

+++ MAVGCL/src/com/comino/flight/parameter/ParamUtils.java
@@ -43,22 +43,22 @@ public class ParamUtils {
 	// WEIRD: Representation of int32 in the float field ....
 
 
-	public static float paramToVal (int type, float val)
+	public static double paramToVal (int type, float val)
 	{
 		if(type == MAV_PARAM_TYPE.MAV_PARAM_TYPE_INT32) {
-			return (float) (ByteBuffer.allocate(4).putFloat(0,val).getInt(0));
+			return (double) (ByteBuffer.allocate(4).putFloat(0,val).getInt(0));
 		}
 		else
 			return val ;
 	}
 
-	public static float valToParam (int type, float val)
+	public static float valToParam (int type, double val)
 	{
 		if(type == MAV_PARAM_TYPE.MAV_PARAM_TYPE_INT32) {
 			return (float) (ByteBuffer.allocate(4).putInt(0,(int)val).getFloat(0));
 		}
 		else
-			return val ;
+			return (float)val ;
 	}
 
 }
